optimize speed: separate maps from triggers
git-svn-id: https://svn.ego-gw.it/svn/advsw/Omicron/trunk@59432 47e613b6-93f3-4244-b775-e06b1756610f

+++ src/Omicron.cc
@@ -529,9 +529,9 @@ bool Omicron::Project(void){
       if(fVerbosity>2) cout<<"\t\t- save triggers"<<endl;
       if(!tile->SaveTriggers(triggers[chanindex],
 			     fSNRThreshold,
-			     dataseq->GetCurrentOverlapDuration()-dataseq->GetOverlapDuration()/2,
-			     dataseq->GetOverlapDuration()/2,
-			     dataseq->GetSegmentTimeStart(s)+dataseq->GetSegmentDuration()/2)
+			     (double)(dataseq->GetCurrentOverlapDuration()-dataseq->GetOverlapDuration()/2),
+			     (double)(dataseq->GetOverlapDuration()/2),
+			     (double)(dataseq->GetSegmentTimeStart(s)+dataseq->GetSegmentDuration()/2))
 	 ) return false;
     }
